% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georgia_dph_plot.R
\name{make_georgia_dph_plot}
\alias{make_georgia_dph_plot}
\title{Make a fixed version of the plot made by Georgia's DPH}
\usage{
make_georgia_dph_plot(georgia_data)
}
\arguments{
\item{georgia_data}{The georgia_covid data set from the buad312data library. Expects the dataframe to have columns county, date, cases}
}
\value{
ggplot object with the bar plot
}
\description{
Make a fixed version of the plot made by Georgia's DPH
}
