% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightingale.R
\docType{data}
\name{nightingale}
\alias{nightingale}
\title{Dataset for Florence Nightingale's famous polar area diagram.}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{date}{The month of the reported data}
\item{army_size}{The number of people in the British army during this month}
\item{number_of_deaths}{average annual mortality per 1000 for each cause}
\item{cause_of_death}{reported cause of death. (disease, wounds, other)}
}
}
\source{
\url{https://github.com/datasets-io/nightingales-rose}
}
\usage{
nightingale
}
\description{
A data set containing the causes of mortality in the British army during the Crimean War
}
\keyword{datasets}
