% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{LA Metro bike share stations}
\format{
A data frame with 245 rows and 5 variables:
\describe{
\item{Station_ID}{Unique integer that identifies the station (this is the same ID used in the Trips and Station Status data)}
\item{Station_Name}{The public name of the station. "Virtual Station" is used by staff to check in or check out a bike remotely for a special event or in a situation in which a bike could not otherwise be checked in or out to a station.}
\item{Go_live_date}{The date that the station was first available}
\item{Region}{he municipality or area where a station is located, includes DTLA (Downtown LA), Pasadena, Port of LA, Venice}
\item{Status}{"Active" for stations available or "Inactive" for stations that are not available as of the latest update}
}
}
\source{
\url{https://bikeshare.metro.net/about/data/}
}
\usage{
stations
}
\description{
A dataset containing LA Metro bike share stations
}
\keyword{datasets}
